+++ pkg/sql/BUILD.bazel
@@ -416,6 +416,9 @@ go_library(
         "//pkg/sql/sqlutil",
         "//pkg/sql/stats",
         "//pkg/sql/stmtdiagnostics",
+        "//pkg/sql/storageparam",
+        "//pkg/sql/storageparam/indexstorageparam",
+        "//pkg/sql/storageparam/tablestorageparam",
         "//pkg/sql/types",
         "//pkg/sql/vtable",
         "//pkg/storage/enginepb",
